requirements.txt
openai
tiktoken
chromadb
pandas
python-dotenv
ipykernel
langchain
streamlit
pysqlite3-binary
kor
pandas